+++ packages/babel-helpers/src/helpers.js
@@ -903,7 +903,7 @@ helpers.iterableToArrayLimit = helper("7.0.0-beta.0")`
     // _i = _iterator
     // _s = _step
 
-    var _i = arr && (typeof Symbol !== "undefined" && arr[Symbol.iterator] || arr["@@iterator"]);
+    var _i = arr == null ? null : (typeof Symbol !== "undefined" && arr[Symbol.iterator] || arr["@@iterator"]);
     if (_i == null) return;
 
     var _arr = [];