+++ packages/babel-parser/src/parser/util.js
@@ -14,6 +14,7 @@ import ProductionParameterHandler, {
   PARAM,
 } from "../util/production-parameter";
 import { Errors, type ErrorTemplate, ErrorCodes } from "./error";
+import type { ParsingError } from "./error";
 /*::
 import type ScopeHandler from "../util/scope";
 */
@@ -175,6 +176,7 @@ export default class UtilParser extends Tokenizer {
         template: `Unexpected token, expected "${messageOrType.label}"`,
       };
     }
+
     /* eslint-disable @babel/development-internal/dry-error-messages */
     throw this.raise(pos != null ? pos : this.state.start, messageOrType);
     /* eslint-enable @babel/development-internal/dry-error-messages */
@@ -211,7 +213,7 @@ export default class UtilParser extends Tokenizer {
     oldState: State = this.state.clone(),
   ):
     | TryParse<T, null, false, false, null>
-    | TryParse<T | null, SyntaxError, boolean, false, State>
+    | TryParse<T | null, ParsingError, boolean, false, State>
     | TryParse<T | null, null, false, true, State> {
     const abortSignal: { node: T | null } = { node: null };
     try {
@@ -228,7 +230,7 @@ export default class UtilParser extends Tokenizer {
         this.state.tokensLength = failState.tokensLength;
         return {
           node,
-          error: (failState.errors[oldState.errors.length]: SyntaxError),
+          error: (failState.errors[oldState.errors.length]: ParsingError),
           thrown: false,
           aborted: false,
           failState,
@@ -267,13 +269,21 @@ export default class UtilParser extends Tokenizer {
     andThrow: boolean,
   ) {
     if (!refExpressionErrors) return false;
-    const { shorthandAssign, doubleProto } = refExpressionErrors;
-    if (!andThrow) return shorthandAssign >= 0 || doubleProto >= 0;
+    const { shorthandAssign, doubleProto, optionalParameters } =
+      refExpressionErrors;
+    if (!andThrow) {
+      return (
+        shorthandAssign >= 0 || doubleProto >= 0 || optionalParameters >= 0
+      );
+    }
     if (shorthandAssign >= 0) {
       this.unexpected(shorthandAssign);
     }
     if (doubleProto >= 0) {
       this.raise(doubleProto, Errors.DuplicateProto);
+    }
+    if (optionalParameters >= 0) {
+      this.unexpected(optionalParameters);
     }
   }
 
@@ -394,17 +404,19 @@ export default class UtilParser extends Tokenizer {
 }
 
 /**
- * The ExpressionErrors is a context struct used to track
- * - **shorthandAssign**: track initializer `=` position when parsing ambiguous
- *   patterns. When we are sure the parsed pattern is a RHS, which means it is
- *   not a pattern, we will throw on this position on invalid assign syntax,
- *   otherwise it will be reset to -1
- * - **doubleProto**: track the duplicate `__proto__` key position when parsing
- *   ambiguous object patterns. When we are sure the parsed pattern is a RHS,
- *   which means it is an object literal, we will throw on this position for
- *   __proto__ redefinition, otherwise it will be reset to -1
+ * The ExpressionErrors is a context struct used to track ambiguous patterns
+ * When we are sure the parsed pattern is a RHS, which means it is not a pattern,
+ * we will throw on this position on invalid assign syntax, otherwise it will be reset to -1
+ *
+ * Types of ExpressionErrors:
+ *
+ * - **shorthandAssign**: track initializer `=` position
+ * - **doubleProto**: track the duplicate `__proto__` key position
+ * - **optionalParameters**: track the optional paramter (`?`).
+ * It's only used by typescript and flow plugins
  */
 export class ExpressionErrors {
   shorthandAssign = -1;
   doubleProto = -1;
+  optionalParameters = -1;
 }