+++ packages/babel-parser/test/fixtures/typescript/class/constructor-with-invalid-order-modifiers-3/input.ts
@@ -1,0 +1,3 @@
+class D extends B {
+  constructor(override readonly public foo: string) {}
+}