+++ packages/babel-plugin-transform-spread/test/fixtures/spread/new-expression-babel-7/output.js
@@ -1,0 +1,3 @@
+babelHelpers.construct(Numbers, babelHelpers.toConsumableArray(nums));
+babelHelpers.construct(Numbers, [1].concat(babelHelpers.toConsumableArray(nums)));
+babelHelpers.construct(Numbers, babelHelpers.arrayWithoutHoles([1,, 3]));