+++ packages/babel-plugin-transform-spread/test/fixtures/spread/new-expression/output.js
@@ -1,2 +1,3 @@
 babelHelpers.construct(Numbers, babelHelpers.toConsumableArray(nums));
 babelHelpers.construct(Numbers, [1].concat(babelHelpers.toConsumableArray(nums)));
+babelHelpers.construct(Numbers, babelHelpers.arrayLikeToArray([1,, 3]));