+++ packages/babel-plugin-transform-typescript/src/namespace.ts
@@ -68,9 +68,10 @@ function handleVariableDeclaration(
   }
   const { declarations } = node;
   if (
-    declarations.every((declarator): declarator is t.VariableDeclarator & {
-      id: t.Identifier;
-    } => t.isIdentifier(declarator.id))
+    declarations.every(
+      (declarator): declarator is t.VariableDeclarator & { id: t.Identifier } =>
+        t.isIdentifier(declarator.id),
+    )
   ) {
     // `export const a = 1` transforms to `const a = N.a = 1`, the output
     // is smaller than `const a = 1; N.a = a`;